+++ cypress/support/component.ts
@@ -1,5 +1,4 @@
 import "./commands";
-import "cypress-mochawesome-reporter/register";
 
 /* returning false here prevents Cypress from failing the test */
 Cypress.on("uncaught:exception", () => false);