+++ src/components/checkbox/checkbox.stories.tsx
@@ -18,6 +18,7 @@ export const Default: ComponentStory<typeof Checkbox> = () => {
 export const Sizes: ComponentStory<typeof Checkbox> = () => (
   <>
     <Checkbox
+      mb={2}
       label="Small"
       key="checkbox-small"
       name="checkbox-small"
@@ -69,9 +70,8 @@ export const WithLabelHelp: ComponentStory<typeof Checkbox> = () => (
 
 export const WithCustomLabelWidth: ComponentStory<typeof Checkbox> = () => (
   <Checkbox
-    label="With custom labelWidth and label aligned to right"
+    label="With custom labelWidth"
     labelWidth={100}
-    labelAlign="right"
     name="checkbox-custom-label"
   />
 );