+++ src/components/multi-action-button/multi-action-button.style.ts
@@ -100,14 +100,6 @@ const StyledButtonChildrenContainer = styled.div<StyledButtonChildrenContainerPr
       text-align: ${align};
       z-index: ${theme.zIndex.overlay};
 
-      /* Styling for Safari. */
-      @media not all and (min-resolution: 0.001dpcm) {
-        @supports (-webkit-appearance: none) and (stroke-color: transparent) {
-          display: -webkit-box;
-          justify-content: ${align === "right" ? `flex-end` : `flex-start`};
-        }
-      }
-
       & + & {
         margin-top: 3px;
       }