+++ src/components/popover-container/popover-container-test.stories.tsx
@@ -1,14 +1,17 @@
 import React, { useState } from "react";
 import Button from "../button";
 import Box from "../box";
-import PopoverContainer, {
-  PopoverContainerProps,
-} from "./popover-container.component";
+import PopoverContainer from "./popover-container.component";
 import { Select, Option } from "../select";
+import { Menu, MenuItem } from "../menu";
+import Heading from "../heading";
+import Typography from "../typography";
+import IconButton from "../icon-button";
+import Icon from "../icon";
 
 export default {
   title: "Popover Container/Test",
-  includeStories: ["Default", "WithSelect", "InAScrollableBlock"],
+  includeStories: ["Default", "WithSelect", "InAScrollableBlock", "InSideMenu"],
   parameters: {
     info: { disable: true },
     chromatic: {
@@ -51,54 +54,6 @@ Default.story = {
   },
 };
 
-export const PopoverContainerComponent = (
-  props: Partial<PopoverContainerProps>
-) => {
-  const [isOpen, setIsOpen] = useState(true);
-
-  const onOpen = () => setIsOpen(isOpen);
-  const onClose = () => setIsOpen(!isOpen);
-
-  return (
-    <div
-      style={{
-        height: 150,
-        margin: "100px",
-      }}
-    >
-      <PopoverContainer
-        title="Cypress is awesome"
-        containerAriaLabel="popover-container"
-        openButtonAriaLabel="open"
-        open={isOpen}
-        onOpen={onOpen}
-        onClose={onClose}
-        {...props}
-      >
-        Contents
-      </PopoverContainer>
-    </div>
-  );
-};
-
-export const PopoverContainerWithSelect = () => {
-  return (
-    <div style={{ height: 100 }}>
-      <PopoverContainer
-        containerAriaLabel="popover-container"
-        openButtonAriaLabel="open"
-        title="select example"
-      >
-        <Select label="my select">
-          <Option value="red" text="red" />
-          <Option value="green" text="green" />
-          <Option value="blue" text="blue" />
-        </Select>
-      </PopoverContainer>
-    </div>
-  );
-};
-
 export const InAScrollableBlock = () => {
   return (
     <Box>
@@ -136,5 +91,46 @@ export const InAScrollableBlock = () => {
         </Box>
       </Box>
     </Box>
+  );
+};
+
+export const InSideMenu = () => {
+  const [open, setOpen] = useState(false);
+  return (
+    <Menu menuType="black">
+      <MenuItem flex="0 0 auto">
+        <PopoverContainer
+          position="left"
+          shouldCoverButton
+          onOpen={() => setOpen(true)}
+          onClose={() => setOpen(false)}
+          open={open}
+          // eslint-disable-next-line @typescript-eslint/no-unused-vars
+          renderOpenComponent={({ isOpen, ref, onClick }) => (
+            <IconButton aria-label="Notifications" ref={ref} onAction={onClick}>
+              <Icon type="alert" />
+            </IconButton>
+          )}
+          p={0}
+        >
+          <Box mt="-8px" backgroundColor="#f2f5f6ff">
+            <Heading
+              title={
+                <Box mt={2} ml={2}>
+                  Notifications
+                </Box>
+              }
+              subheader={<Typography ml={2}>99 red balloons</Typography>}
+            />
+          </Box>
+          <Box display="flex" justifyContent="space-between" p={2}>
+            <Button size="small">Mark all as read</Button>
+            <Button buttonType="primary" size="small">
+              View all notifications
+            </Button>
+          </Box>
+        </PopoverContainer>
+      </MenuItem>
+    </Menu>
   );
 };