+++ src/components/select/option/option.component.tsx
@@ -1,5 +1,6 @@
-import React, { useContext } from "react";
-import { TagProps } from "__internal__/utils/helpers/tags";
+import React, { useContext, useRef } from "react";
+import guid from "../../../__internal__/utils/helpers/guid";
+import { TagProps } from "../../../__internal__/utils/helpers/tags";
 import StyledOption from "./option.style";
 import SelectListContext from "../__internal__/select-list-context";
 
@@ -66,6 +67,7 @@ const Option = React.forwardRef(
   ) => {
     const selectListContext = useContext(SelectListContext);
     let isSelected = selectListContext.currentOptionsListIndex === index;
+    const internalIdRef = useRef(id || guid());
 
     if (selectListContext.multiselectValues) {
       isSelected = selectListContext.multiselectValues.includes(value);
@@ -76,7 +78,7 @@ const Option = React.forwardRef(
         return;
       }
       if (!onClick) {
-        onSelect?.({ text, value, id });
+        onSelect?.({ text, value, id: internalIdRef.current });
       } else {
         onSelect?.();
         onClick(value);
@@ -85,7 +87,7 @@ const Option = React.forwardRef(
 
     return (
       <StyledOption
-        id={id}
+        id={internalIdRef.current}
         ref={ref}
         aria-selected={isSelected}
         aria-disabled={disabled}