+++ src/components/step-flow/index.ts
@@ -1,0 +1,2 @@
+export { default as StepFlow } from "./step-flow.component";
+export type { StepFlowProps } from "./step-flow.component";